refactor(types): Start a folder for utility types PE-635
