+++ src/community/contract_types.ts
@@ -1,4 +1,4 @@
-import { ArweaveAddress } from '../arweave_address';
+import { ArweaveAddress } from '../types/arweave_address';
 
 export type CommunityTipPercentage = number;
