+++ src/wallet.ts
@@ -15,7 +15,7 @@ import {
 	RewardSettings
 } from './types';
 import { CreateTransactionInterface } from 'arweave/node/common';
-import { ArweaveAddress } from './arweave_address';
+import { ArweaveAddress } from './types/arweave_address';
 
 export type ARTransferResult = {
 	trxID: TransactionID;
